+++ sys/arch/hp300/hp300/mem.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: mem.c,v 1.20 1998/08/20 08:33:44 kleink Exp $	*/
+/*	$NetBSD: mem.c,v 1.21 1998/11/19 15:38:22 mrg Exp $	*/
 
 /*
  * Copyright (c) 1988 University of Utah.
@@ -237,7 +237,7 @@ mmmmap(dev, off, prot)
 	/*
 	 * Allow access only in RAM.
 	 */
-	if ((unsigned)off < lowram || (unsigned)off >= 0xFFFFFFFC)
+	if ((u_int)off < lowram || (u_int)off >= 0xFFFFFFFC)
 		return (-1);
-	return (m68k_btop(off));
+	return (m68k_btop((u_int)off));
 }
