+++ sys/arch/mac68k/obio/asc.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: asc.c,v 1.28 1998/08/15 07:42:50 scottr Exp $	*/
+/*	$NetBSD: asc.c,v 1.29 1998/11/19 15:38:23 mrg Exp $	*/
 
 /*
  * Copyright (C) 1997 Scott Reynolds
@@ -288,7 +288,7 @@ ascmmap(dev, off, prot)
 	vm_offset_t pa;
 
 	sc = asc_cd.cd_devs[unit];
-	if (off < MAC68K_ASC_LEN) {
+	if ((u_int)off < MAC68K_ASC_LEN) {
 		pa = pmap_extract(pmap_kernel(), (vm_offset_t)sc->sc_handle);
 		return m68k_btop(pa + off);
 	}
