+++ sys/arch/mvme68k/mvme68k/mem.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: mem.c,v 1.5 1998/08/22 10:55:35 scw Exp $	*/
+/*	$NetBSD: mem.c,v 1.6 1998/11/19 15:38:23 mrg Exp $	*/
 
 /*
  * Copyright (c) 1988 University of Utah.
@@ -228,7 +228,7 @@ mmmmap(dev, off, prot)
 	 * XXX could be extended to allow access to IO space but must
 	 * be very careful.
 	 */
-	if ((unsigned)off < lowram || (unsigned)off >= 0xFFFFFFFC)
+	if ((u_int)off < lowram || (u_int)off >= 0xFFFFFFFC)
 		return (-1);
-	return (m68k_btop(off));
+	return (m68k_btop((u_int)off));
 }
